fix(popovermenu/canvas): standarised controls and switched to ts file
